refactor(cli): use env! to get version and name
Instead of having to update manually the version that is shown by the CLI
we get this from the Cargo.toml file using `env!` macro. Now whenever
version number is needed in the program it is going to be extracted from
Cargo.toml.

The same can be for package authors, description etc.